Decouple signup rewards from notification routes
